refactor: 학생회 컨트롤러 기능별 분리 및 파싱 로직 제거
